make use of pimpl idiom
