+++ include/ringmesh/geomodel/geomodel_builder.h
@@ -36,6 +36,7 @@
 #pragma once
 
 #include <ringmesh/basic/common.h>
+#include <ringmesh/basic/pimpl.h>
 
 #include <ringmesh/geomodel/geomodel_builder_geology.h>
 #include <ringmesh/geomodel/geomodel_builder_geometry.h>
@@ -159,39 +160,12 @@ namespace RINGMesh
     public:
         explicit GeoModelBuilder( GeoModel2D& geomodel );
 
+        ~GeoModelBuilder();
+
         void build_surfaces_from_corners_and_lines();
 
     private:
-        struct OrientedLine
-        {
-            OrientedLine( index_t line_index, bool line_side )
-                : index( line_index ), side( line_side )
-            {
-            }
-            index_t index;
-            bool side;
-        };
-
-        void find_surfaces_boundary_lines(
-            std::vector< index_t >& line_2_surface,
-            std::vector< std::vector< OrientedLine > >&
-                surface_boundary_lines );
-
-        void check_internal_floating_lines(
-            const std::vector< index_t >& line_2_surface,
-            const index_t nb_found_surfaces );
-
-        void build_surface_polygons(
-            const std::vector< std::vector< OrientedLine > >&
-                surface_boundary_lines );
-
-        void find_exterior_and_remove_it(
-            std::vector< std::vector< OrientedLine > >&
-                surface_boundary_lines );
-
-        void set_surface_line_boundary_relationships(
-            const std::vector< std::vector< OrientedLine > >&
-                surface_boundary_lines );
+        IMPLEMENTATION_MEMBER( impl_ );
     };
 
     template <>
